Small changes, mostly docstrings
